requirements.txt
fastapi
fastapi_jwt_auth
bcrypt
python-jose
python-multipart
aiofiles
requests
pytest
passlib
psycopg2
mypy
peewee
uvicorn
uvicorn[standard]
importlib-metadata
aiohttp
contextvars
google-cloud-storage